# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRC API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "api/"
        }
    ],
    "paths": {
...
        "/submit-scouting-report": {
            "post": {
                "tags": [
                    "Scouts"
                ],
                "summary": "Submit player's scouting report",
                "description": "submit report",
                "operationId": "submitScoutingReport",
                "requestBody": {
                    "description": "pass player_id, game, skills, longRangePotential, scoutComment, recommendation, published, scoutRequestId",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "game",
                                    "longRangePotential",
                                    "scoutComment",
                                    "recommendation",
                                    "scoutRequestId"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "scout_request_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "game": {
                                        "type": "string",
                                        "example": "hockey"
                                    },
                                    "skills": {
                                        "type": "json",
                                        "example": "{}"
                                    },
                                    "long_range_potential": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "scout_comment": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "recommendation": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "published": {
                                        "type": "boolean",
                                        "example": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Assessment report saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/v2/submit-scouting-report": {
            "post": {
                "tags": [
                    "Scouts"
                ],
                "summary": "Submit player's scouting report",
                "description": "submit report",
                "operationId": "submitScoutingReportV2",
                "requestBody": {
                    "description": "pass player_id, game, skills, longRangePotential, scoutComment, recommendation, published, scoutRequestId, modified_skills",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "game",
                                    "longRangePotential",
                                    "scoutComment",
                                    "recommendation",
                                    "scoutRequestId",
                                    "modified_skills"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "scout_request_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "game": {
                                        "type": "string",
                                        "example": "hockey"
                                    },
                                    "skills": {
                                        "type": "string",
                                        "example": "[]"
                                    },
                                    "modified_skills": {
                                        "type": "string",
                                        "example": "[]"
                                    },
                                    "long_range_potential": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "scout_comment": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "recommendation": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "published": {
                                        "type": "boolean",
                                        "example": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Assessment report saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/v3/submit-scouting-report": {
            "post": {
                "tags": [
                    "Scouts"
                ],
                "summary": "Submit player's scouting report",
                "description": "submit report",
                "operationId": "submitScoutingReportV3",
                "requestBody": {
                    "description": "pass player_id, game, skills, longRangePotential, scoutComment, recommendation, published, scoutRequestId, modified_skills",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "game",
                                    "longRangePotential",
                                    "scoutComment",
                                    "recommendation",
                                    "scoutRequestId",
                                    "modified_skills"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "scout_request_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "game": {
                                        "type": "string",
                                        "example": "hockey"
                                    },
                                    "skills": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "modified_skills": {
                                        "type": "string",
                                        "example": "[]"
                                    },
                                    "long_range_potential": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "scout_comment": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "recommendation": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "published": {
                                        "type": "boolean",
                                        "example": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Assessment report saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "Authentication Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}